Initial try
When I first started I was writing a new base
class for the wrapper class and a new version
of the product_id class. At some time I discussed
my progress with Chris and he really wanted to
actually have the wrapper class inherit from
the existing event_base class. So I am ending this
branch and starting a new one pursuing that course.
Possibly I can come back to this later and use some
of this code. I think it is better but it is not
a minimal change.