Implement wrapper class
The wrapper template is the top level class
that will be written to output files. It will
contain an instance of its template parameter
which will be the data format type that users
are interested in. Wrapper inherits from
event_base instead of having all the user
data types inherit from event_base.

I added tests of 3 types to test the wrapper.
A vector<hit1> type, a fundamental type
int, and a map<int, double>. If this code
is used, I expect the hit1 class to be renamed
hit and replace the existing hit class.

There is test code to read and write these
types in a C++ program and also in a pyroot
script.

The data_type template function is moved into
a separate file from storage_manager so that
wrapper does not have to include the
storage_manager header file.