Add wrapper and hit1 classes
This is prototype code for adding a
wrapper template class each of whose
instances will contain a fundamental
data type. Wrapper inherits from
event_base instead of having the
fundamental data types have a type
that inherits from a vector<type>
and event_base.

If we proceed in the direction of using
the wrapper class I expect that a future
step would be to make the hit1 class
replace the hit class instead of being an
additional class. And a similar change
would be made to all the data format
classes.