Minor fixes so tests will work on ROOT 5
Commented out some tests of std::map with pyROOT
because they do not work using ROOT 5.

Fixed the LinkDef file to create a dictionary for
std::pair<const int, double> instead of
std:pair<int, double>. It did not matter for ROOT 6
for some reason, but was causing a seg fault for
ROOT 5. This is the correct value_type of the map.

Fix a preexisting bug in one of the tests which
produced a spurious error message.