Read larsoft data product
This contains the changes to read the larsoft
data product vector<Hit> with larlite directly
linking to the larsoft libraries.

The C++ executables in core/DataFormats/bin work
OK. The pyROOT does not and the python scripts are
in a broken unfinished state where I was trying to
debug the problem.  We believe the problem is that
larsoft builds dictionaries with only the IO
related functions and nothing else. Larlite
needs more than that. Further there is a bug
in the flag needed to direct cmake to build
complete dictionaries. It does not seem to
work. This problem should be resolved when we
move ROOT 6 so we decided not to waste any
more time on it in ROOT 5.

This is just a test prototype. I expect it would need
more polishing before committing to the main
larlite branch.