Demonstrator for use case 1
This code implements the demonstrator for use case 1
of the LArSoft/LArLite integration plan. With it one can
run a LArLite executable that writes a LArLite format
file containing a product defined in a LArSoft library.
And also another LArLite executable that will read
the LArSoft product from that file.

It includes this and other things.

- a larlite wrapper class that holds the product when it is
written to the file and the dictionary for the wrapped
LArSoft product.

- definitions of some other products. A class like the
larlite hit class that is written into the file in the
wrapper.

- Also an int and a map<int, double> are written into
the output file in the wrapper.

- It contains the necessary additions to the GNUmakefiles
to find the LArSoft header files and the LArSoft libraries.
It expects that the test LArSoft libraries Patrick Gartung
implemented have been built and setup before this is built or
run.

- It contains the changes in storage_manager necessary for
this to work.

- It contains pyROOT scripts that will read and write the
same file as the built executable.

- It also has fixes unrelated to its primary purpose, because
I encountered problems when I tried to build the head version
of larlite with ROOT 6 version Patrick used without any
modifications to the larlite code. So I fixed the problems or
commented out the offending code where it was unrelated.